# docs/swagger.yaml
definitions:
...
  models.ItemRequest:
    properties:
      descricao:
        type: string
      extra:
        type: boolean
      id:
        type: integer
      preco:
        type: number
      tipo:
        type: string
    required:
    - descricao
    - extra
    - id
    - preco
    - tipo
    type: object
  models.ItemResponse:
    properties:
      descricao:
        type: string
      extra:
        type: boolean
      id:
        type: integer
      preco:
        type: number
      tipo:
        type: string
    type: object
  models.ItemUpdateRequest:
    properties:
      descricao:
        type: string
      extra:
        type: boolean
      preco:
        type: number
    required:
    - descricao
    - extra
    - preco
    type: object
...
info:
  contact: {}
paths:
...
  /itens:
    post:
      consumes:
      - application/json
      description: Cria um novo item (bebida ou ingrediente) com os dados fornecidos
      parameters:
      - description: Dados do Item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.ItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ItemResponse'
        "400":
          description: Erro na validação dos dados
          schema:
            type: string
      summary: Cria um novo item
      tags:
      - itens
  /itens/{codigo}:
    delete:
      consumes:
      - application/json
      description: Remove um item (bebida ou ingrediente) existente pelo código
      parameters:
      - description: Código do item
        in: path
        name: codigo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Item removido com sucesso
          schema:
            type: string
        "400":
          description: Código inválido
          schema:
            type: string
        "404":
          description: Item não encontrado
          schema:
            type: string
      summary: Deleta um item existente
      tags:
      - itens
    get:
      consumes:
      - application/json
      description: Retorna um item específico (bebida ou ingrediente) buscando por
        código
      parameters:
      - description: Código do item
        in: path
        name: codigo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ItemResponse'
        "400":
          description: Código inválido
          schema:
            type: string
        "404":
          description: Item não encontrado
          schema:
            type: string
      summary: Busca um item por código
      tags:
      - itens
    put:
      consumes:
      - application/json
      description: Atualiza um item (bebida ou ingrediente) existente
      parameters:
      - description: Código do item
        in: path
        name: codigo
        required: true
        type: string
      - description: Dados do Item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.ItemUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ItemResponse'
        "400":
          description: Código inválido
          schema:
            type: string
        "404":
          description: Item não encontrado
          schema:
            type: string
      summary: Atualiza um item existente
      tags:
      - itens
  /itens/bebidas:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de todas as bebidas disponíveis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ItemResponse'
            type: array
        "404":
          description: Nenhuma bebida encontrada
          schema:
            type: string
      summary: Lista todas as bebidas
      tags:
      - itens
  /itens/ingredientes:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de todos os ingredientes disponíveis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ItemResponse'
            type: array
        "404":
          description: Nenhum ingrediente encontrado
          schema:
            type: string
      summary: Lista todos os ingredientes
      tags:
      - itens
  /itens/todos:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de todos os itens (bebidas e ingredientes)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ItemResponse'
            type: array
        "404":
          description: Nenhum item encontrado
          schema:
            type: string
      summary: Lista todos os itens
      tags:
      - itens
...
swagger: "2.0"